separate saveToml
